+++ x/denommetadata/types/expected_keepers.go
@@ -3,19 +3,14 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/cosmos/ibc-go/v6/modules/core/exported"
-	rollapptypes "github.com/dymensionxyz/dymension/v3/x/rollapp/types"
+	transfertypes "github.com/cosmos/ibc-go/v6/modules/apps/transfer/types"
 	tmbytes "github.com/tendermint/tendermint/libs/bytes"
 )
 
 // TransferKeeper defines the expected transfer keeper
 type TransferKeeper interface {
 	HasDenomTrace(ctx sdk.Context, denomTraceHash tmbytes.HexBytes) bool
-}
-
-// ChannelKeeper defines the expected IBC channel keeper
-type ChannelKeeper interface {
-	GetChannelClientState(ctx sdk.Context, portID, channelID string) (string, exported.ClientState, error)
+	SetDenomTrace(ctx sdk.Context, denomTrace transfertypes.DenomTrace)
 }
 
 // BankKeeper defines the expected interface needed
@@ -24,8 +19,6 @@ type BankKeeper interface {
 	SetDenomMetaData(ctx sdk.Context, denomMetaData types.Metadata)
 }
 
-type RollappKeeper interface {
-	GetParams(ctx sdk.Context) rollapptypes.Params
-	GetRollapp(ctx sdk.Context, chainID string) (rollapp rollapptypes.Rollapp, found bool)
-	RegisterDenomMetadata(ctx sdk.Context, rollapp rollapptypes.Rollapp, ibcBaseDenom, baseDenom string)
+type DenomMetadataKeeper interface {
+	CreateDenomMetadata(ctx sdk.Context, metadata types.Metadata) error
 }